requirements.txt
python-dotenv==1.0.1
openai==1.61.1
pandas==2.2.3
scikit-learn==1.6.1
rich==13.9.4
tqdm==4.67.1